test: Adds first convert_to_vector test for n-ary trees to vectors
Prompt: I want you to act as a Ruby developer pair programming with me. You will be the driver and I will be the back seat navigator. I will tell you to write some tests with different expectations. You will use idiomatic rspec to write the tests. Your response will be one unique code block and nothing else. Do not write explanations. I will run the rspec test for you, and provide you with the text output of the command. I will tell you if the test behaves as expected. If the test does not behave as expected, I will tell you to make a change. Your response will always be one unique code block and nothing else. We will proceed in this loop together until we have an rspec test that behaves as I expect it to, then we will switch to implementation mode. When we are in implementation mode, I will ask you to make the test pass. Your response will be one unique code block with the implementation code that will make the test pass. I will run the rspec test with your implementation code, and I will only provide you the output of the rspec test with no other commentary. You will only respond with implementation code in a single code block, nothing else, until the rspec output shows your tests have passed. Then we will switch back to adding more tests. In this way we will pair program. The first test I want you to write is for converting an n-ary tree into a vector.