It thought it had some insight to share and tried updating from that:
The tests are failing because the Tree class does not have the methods to_vector and from_vector, as well as an instance variable @depth. Additionally, the initialize method for the Tree class is not accepting any arguments, which is causing the ArgumentError failures.

To fix these errors, the Tree class needs to have a method to_vector that converts a tree into a vector, a method from_vector that initializes a tree from a vector, and an instance variable @depth that stores the depth of the tree. The initialize method should also be updated to accept a root node as an argument.

Here is an example implementation of the Tree class that should pass the tests: