It is lost:
To fix these test failures, you need to add a root node to your tree in the initialize method. Here is an updated version of the Tree class that should make the tests pass:

Here, we have added a default value of TreeNode.new to the root parameter in the initialize method, so that a root node is always created for the tree. We have also added a from_vector class method that takes a vector and uses it to initialize a new tree instance. This method should be able to correctly initialize a tree and make the tests pass.